gcp-create-account/requirements.txt
google-auth
google-auth-httplib2
google-auth-oauthlib
requests
google-api-python-client